Refactor CanvasPage project load state overlay
